pyproject.toml
[project]
name = "TAER_Core"
version = "0.0.1"
description = "Core modules for TAER (Test AER based sensors)."
authors = [
    {name = "Rafael de la Rosa", email = "rafodelarosa.og@gmail.com"},
    {name = "Ruben Gomez Merchan", email = "rgmerchan@us.es"},
]
requires-python = ">=3.10.5"
dependencies = [
    "numpy==1.24.0",
    "opencv-python",
    "GitPython==3.1.43",
    "matplotlib==3.9.1",
    "PyYAML==6.0.1",
    "wxPython==4.2.0",
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
